feat: finalize API
