+++ README.md
@@ -2,6 +2,18 @@
 
 Muh (say: Moo!) stands for "mustached hypertext" and is the template language that runs [Sissi](https://sissi.js.org), the static site generator. It is put into a single package so it can be re-used in other projects as well.
 
+## Flavors
+
+Muh comes in 2 flavors. Both of them live in the `@sissijs/muh` package.
+
+`@sissijs/muh` includes everything (a full-fledged `processTemplateFile` function that supports SmolYAML/JSON frontmatter, layout files, basic preprocessing such as (a minimal) markdown, handling css imports and resolving `<html-include src="">` directives). In roughly 9KB of JavaScript
+
+`@sissijs/muh/template` is just the `template` function that implements the mustache syntax. It also provides the built-in filters for formatting dates/numbers, and built-in helper functions. The helper functions provided are small wrappers around fetch, namely `fetchText()` and `fetchJson()`. All that is contained in roughly 3KB of JavaScript.
+
+Sizes are measured with minification and without compression.
+
+TypeScript declarations and API documentation are autogenerated from JSDoc.
+
 ## Syntax
 
 The syntax is `{{ moo }}` where `moo` can be any arbitrary JavaScript. JavaScript is evaluated in a safe context. The only globals it can access are the ones you provide, plus a few built-ins.
@@ -34,8 +46,20 @@ There are a few built-in helper functions:
 {{ fetchJson('https://yesno.wtf/api') | async | json }}
 {{ fetchText('https://some-html.api/api/weather') | async }}
 
-{{ include('_includes/partial.html', {title: 'some additional data'}) }} /* TODO */
+{{ include('partial.html', {title: 'some additional data'}) }}
 ```
+
+## Additional Syntaxes
+
+HTML-files can use `<html-include src="partial.html" title="some additional data"/>` directives.
+Additional attributes are passed to the include's data.
+
+There is a basic markdown processor for markdown files involved.
+Markdown files also interprete the `<html-include>` directive.
+
+Also, CSS import directives are resolved.
+
+All file types mentioned above can also work with the mustache syntax.
 
 ## Usage
