+++ package.json
@@ -1,22 +1,38 @@
 {
   "name": "@sissijs/muh",
+  "description": "Muh (say: Moo!) stands for 'mustached hypertext' and is a template language for static site generators",
   "version": "1.0.0",
   "type": "module",
   "main": "src/muh.js",
+  "exports": {
+    ".": "./dist/muh.esm.js",
+    "./template": "./dist/template.esm.js"
+  },
+  "files": [
+    "dist/",
+    "README.md",
+    "LICENSE"
+  ],
   "engines": {
-    "node": ">=20"
+    "node": ">=22"
   },
   "scripts": {
     "lint": "eslint src",
+    "build": "rollup -c rollup.config.js",
+    "types": "tsc -p tsconfig.json",
+    "docs": "typedoc src/muh.js src/typedefs.js src/builtin-filters.js src/builtin-helpers.js",
     "test": "node --test"
   },
   "keywords": [],
-  "author": "",
-  "license": "ISC",
-  "description": "",
+  "author": "Lea Rosema",
+  "license": "MIT",
   "devDependencies": {
     "@eslint/js": "^9.17.0",
+    "@rollup/plugin-terser": "^0.4.4",
     "eslint": "^9.17.0",
-    "globals": "^15.14.0"
+    "globals": "^15.14.0",
+    "rollup": "^4.29.1",
+    "typedoc": "^0.27.6",
+    "typescript": "^5.7.2"
   }
 }
