--- src/index.js
@@ -1,6 +1,0 @@
-import { template, handleTemplateFile } from './muh.js';
-import { smolYAML } from './smolyaml.js';
-import { frontmatter } from './frontmatter.js';
-
-export { template, handleTemplateFile, smolYAML, frontmatter };
-
