+++ src/preprocessors/html.js
@@ -1,18 +1,23 @@
 import { replaceAsync } from "../utils/replace-async.js";
 
-
-
 export const htmlPreprocessor = {
   name: 'html',
   extension: /.html?$/,
+  /**
+   * Process HTML. Resolves `<html-include>` directives via an `include` function provided in data.
+   * 
+   * @param {string} content 
+   * @param {any} data 
+   * @returns processed HTML
+   */
   async process(content, data) {
-    return await replaceAsync(content, /<html-include((\s[a-z]{1,100}?=".{1,1024}?"){1,100})>/gm,
+    return await replaceAsync(content, /<html-include((\s[a-z]{1,100}?=".{1,1024}?"){1,100})\s?\/?>/gm,
       async function (...matches) {
         const match = matches && matches.length >= 2 ? matches[1] : null;
         if (! match) {
           return '';
         }
-        if (typeof data['include'] !== 'function') {
+        if (typeof data?.include !== 'function') {
           return matches[0];
         }
         const params = Object.fromEntries(
@@ -20,7 +25,7 @@ export const htmlPreprocessor = {
           .map(([, key, val]) => [key, val])
           .filter(([key]) => key !== '__proto__')
         );
-        return await data.include(params.src, params);
+        return await data?.include(params.src, params);
       }
     );
   }
