+++ src/typedefs.js
@@ -1,18 +1,25 @@
 /**
  * @typedef Preprocessor
- * @property {string} name
- * @property {string|RegExp} extension
+ * @property {string} name name of the preprocessor.
+ * @property {string|RegExp} extension string or regexp of the extension, including the dot.
  * @property {(content: string, data: any) => Promise<string>} process content processing function
  */
 
 /**
  * @typedef TemplateConfig
  * Configuration object for the template() function.
- * @property {(file: string) => Promise<Buffer|String>} resolve
- * @property {Map<string, Function>} filters
- * @property {string} baseDir default: "."
- * @property {string} includeDir default: "_includes" 
- * @property {string} layoutDir default: "_layouts"
- * @property {boolean} relativeIncludes whether to include relative to the current file
- * @property {Array<Preprocessor>} preprocessors
+ * @property {Map<string, Function>} filters map of additional filters to be used inside the template
  */
+
+/**
+ * @typedef ProcessTemplateFileConfig
+ * Configuration object for the template() function.
+ * @property {(file: string) => Promise<Buffer|String>} [resolve] optional resolver function (defaults to node readFile)
+ * @property {Map<string, Function>} [filters] map of additional filters to be used in the template
+ * @property {string} [baseDir] base directory. default: "."
+ * @property {string} [includeDir="_includes"] subdirectory to look for html partials. default: "_includes"
+ * @property {string} [layoutDir="_layouts"] subdirectory to look for layout files. default: "_layouts"
+ * @property {boolean} [relativeIncludes=false] whether to include relative to the current file. default: false
+ * @property {Array<Preprocessor>} [preprocessors] list of preprocessors. By default: built-ins for html, css, markdown
+ */
+
